+++ src/Io/Internal/WkbTools.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Brick\Geo\Io\Internal;
 
 use Brick\Geo\Exception\GeometryIoException;
+use Brick\Geo\Io\ByteOrder;
 
 /**
  * Tools for WKB classes.
@@ -33,17 +34,17 @@ final readonly class WkbTools
      *
      * @throws GeometryIoException
      */
-    public static function getMachineByteOrder() : WkbByteOrder
+    public static function getMachineByteOrder() : ByteOrder
     {
-        /** @var WkbByteOrder|null $byteOrder */
+        /** @var ByteOrder|null $byteOrder */
         static $byteOrder;
 
         if ($byteOrder === null) {
             self::checkDoubleIs64Bit();
 
             $byteOrder = match (pack('L', 0x61626364)) {
-                'abcd' => WkbByteOrder::BigEndian,
-                'dcba' => WkbByteOrder::LittleEndian,
+                'abcd' => ByteOrder::BigEndian,
+                'dcba' => ByteOrder::LittleEndian,
                 default => throw GeometryIoException::unsupportedEndianness(),
             };
         }