+++ tests/IO/EwkbWriterTest.php
@@ -7,7 +7,7 @@ namespace Brick\Geo\Tests\IO;
 use Brick\Geo\Exception\GeometryIoException;
 use Brick\Geo\Io\EwkbWriter;
 use Brick\Geo\Io\EwktReader;
-use Brick\Geo\Io\Internal\WkbByteOrder;
+use Brick\Geo\Io\ByteOrder;
 use Brick\Geo\Point;
 use PHPUnit\Framework\Attributes\DataProvider;
 
@@ -17,12 +17,12 @@ use PHPUnit\Framework\Attributes\DataProvider;
 class EwkbWriterTest extends EwkbAbstractTestCase
 {
     /**
-     * @param string       $ewkt      The EWKT to read.
-     * @param string       $ewkb      The expected EWKB output, hex-encoded.
-     * @param WkbByteOrder $byteOrder The byte order to use.
+     * @param string    $ewkt      The EWKT to read.
+     * @param string    $ewkb      The expected EWKB output, hex-encoded.
+     * @param ByteOrder $byteOrder The byte order to use.
      */
     #[DataProvider('providerWrite')]
-    public function testWrite(string $ewkt, string $ewkb, WkbByteOrder $byteOrder) : void
+    public function testWrite(string $ewkt, string $ewkb, ByteOrder $byteOrder) : void
     {
         $writer = new EwkbWriter(byteOrder: $byteOrder);
         $reader = new EwktReader();
@@ -36,19 +36,19 @@ class EwkbWriterTest extends EwkbAbstractTestCase
     public static function providerWrite() : \Generator
     {
         foreach (self::providerLittleEndianEwkb() as [$wkt, $ewkb]) {
-            yield [$wkt, $ewkb, WkbByteOrder::LittleEndian];
+            yield [$wkt, $ewkb, ByteOrder::LittleEndian];
         }
 
         foreach (self::providerLittleEndianEwkbWithSrid() as [$wkt, $ewkb]) {
-            yield [$wkt, $ewkb, WkbByteOrder::LittleEndian];
+            yield [$wkt, $ewkb, ByteOrder::LittleEndian];
         }
 
         foreach (self::providerBigEndianEwkb() as [$wkt, $ewkb]) {
-            yield [$wkt, $ewkb, WkbByteOrder::BigEndian];
+            yield [$wkt, $ewkb, ByteOrder::BigEndian];
         }
 
         foreach (self::providerBigEndianEwkbWithSrid() as [$wkt, $ewkb]) {
-            yield [$wkt, $ewkb, WkbByteOrder::BigEndian];
+            yield [$wkt, $ewkb, ByteOrder::BigEndian];
         }
     }
 
@@ -77,7 +77,7 @@ class EwkbWriterTest extends EwkbAbstractTestCase
     }
 
     #[DataProvider('providerWriteEmptyPointWithNanSupport')]
-    public function testWriteEmptyPointWithNanSupport(Point $point, WkbByteOrder $byteOrder, string $expectedHex) : void
+    public function testWriteEmptyPointWithNanSupport(Point $point, ByteOrder $byteOrder, string $expectedHex) : void
     {
         $writer = new EwkbWriter(byteOrder: $byteOrder);
 
@@ -88,14 +88,14 @@ class EwkbWriterTest extends EwkbAbstractTestCase
     public static function providerWriteEmptyPointWithNanSupport() : array
     {
         return [
-            [Point::xyEmpty(), WkbByteOrder::BigEndian, '00000000017ff80000000000007ff8000000000000'],
-            [Point::xyEmpty(), WkbByteOrder::LittleEndian, '0101000000000000000000f87f000000000000f87f'],
-            [Point::xyzEmpty(), WkbByteOrder::BigEndian, '00800000017ff80000000000007ff80000000000007ff8000000000000'],
-            [Point::xyzEmpty(), WkbByteOrder::LittleEndian, '0101000080000000000000f87f000000000000f87f000000000000f87f'],
-            [Point::xymEmpty(), WkbByteOrder::BigEndian, '00400000017ff80000000000007ff80000000000007ff8000000000000'],
-            [Point::xymEmpty(), WkbByteOrder::LittleEndian, '0101000040000000000000f87f000000000000f87f000000000000f87f'],
-            [Point::xyzmEmpty(), WkbByteOrder::BigEndian, '00c00000017ff80000000000007ff80000000000007ff80000000000007ff8000000000000'],
-            [Point::xyzmEmpty(), WkbByteOrder::LittleEndian, '01010000c0000000000000f87f000000000000f87f000000000000f87f000000000000f87f'],
+            [Point::xyEmpty(), ByteOrder::BigEndian, '00000000017ff80000000000007ff8000000000000'],
+            [Point::xyEmpty(), ByteOrder::LittleEndian, '0101000000000000000000f87f000000000000f87f'],
+            [Point::xyzEmpty(), ByteOrder::BigEndian, '00800000017ff80000000000007ff80000000000007ff8000000000000'],
+            [Point::xyzEmpty(), ByteOrder::LittleEndian, '0101000080000000000000f87f000000000000f87f000000000000f87f'],
+            [Point::xymEmpty(), ByteOrder::BigEndian, '00400000017ff80000000000007ff80000000000007ff8000000000000'],
+            [Point::xymEmpty(), ByteOrder::LittleEndian, '0101000040000000000000f87f000000000000f87f000000000000f87f'],
+            [Point::xyzmEmpty(), ByteOrder::BigEndian, '00c00000017ff80000000000007ff80000000000007ff80000000000007ff8000000000000'],
+            [Point::xyzmEmpty(), ByteOrder::LittleEndian, '01010000c0000000000000f87f000000000000f87f000000000000f87f000000000000f87f'],
         ];
     }
 }